move from cra to vite
