+++ client/package.json
@@ -8,25 +8,29 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.23",
-    "@types/node": "^14.17.3",
+    "@types/node": "^18.19.78",
     "@types/react": "^17.0.11",
     "@types/react-dom": "^17.0.7",
+    "@vitejs/plugin-react-swc": "^3.8.0",
     "antd": "^4.16.3",
-    "axios": "^0.21.4",
+    "axios": "^1.8.1",
     "dayjs": "^1.10.5",
     "exif-js": "^2.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
     "sass": "^1.35.1",
     "typescript": "^4.3.3",
+    "vite": "^6.2.0",
+    "vite-plugin-svgr": "^4.3.0",
+    "vite-tsconfig-paths": "^5.1.4",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview",
+    "test": "vitest",
+    "test:coverage": "vitest run --coverage --watch=false"
   },
   "eslintConfig": {
     "extends": [
@@ -46,5 +50,11 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {}
+  "devDependencies": {
+    "@vitest/coverage-v8": "^3.0.7",
+    "jsdom": "^26.0.0",
+    "less": "^4.2.2",
+    "sass-embedded": "^1.85.1",
+    "vitest": "^3.0.7"
+  }
 }
