+++ client/src/App.scss
@@ -1,11 +1,10 @@
-@import "~antd/dist/antd.css";
+@import "antd/dist/antd.css";
 
 .main-container {
   padding: 20px;
   width: 60%; /* update to be adaptive with media queries */
-  @media only screen 
-    and (orientation: portrait) { 
-      width: 100%;
+  @media only screen and (orientation: portrait) {
+    width: 100%;
   }
   margin: 0 auto;
