+++ client/src/components/PhotoViewer.jsx
