+++ client/vite.config.ts
@@ -1,0 +1,28 @@
+import { defineConfig } from 'vitest/config'
+import react from '@vitejs/plugin-react-swc'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+    base: '/',
+    plugins: [react()],
+    test: {
+        globals: true,
+        environment: 'jsdom',
+        setupFiles: './src/setupTests.ts',
+        css: true,
+        reporters: ['verbose'],
+        coverage: {
+            reporter: ['text', 'json', 'html'],
+            include: ['src/**/*'],
+            exclude: [],
+        }
+    },
+    css: {
+        preprocessorOptions: {
+            less: {
+                javascriptEnabled: true,
+                additionalData: '@root-entry-name: default;',
+            },
+        },
+    },
+})
